improvements

+++ src/year_2022/day_04.rs
@@ -5,7 +5,7 @@ pub fn part_one(input: &str) -> Solution {
     let fully_contains_count = input
         .lines()
         .map(|l| l.split(","))
-        .map(|s| s.map(|i| i.split("-").map(|j| j.parse::<i32>().unwrap()).collect_tuple::<(_, _)>().unwrap()).collect_vec())
+        .map(|s| s.map(|i| i.split("-").map(|j| j.parse::<u16>().unwrap()).collect_tuple::<(_, _)>().unwrap()).collect_vec())
         .filter(|r| fully_contains(r[0], r[1]))
         .count();
 
@@ -16,19 +16,18 @@ pub fn part_two(input: &str) -> Solution {
     let overlaps_count = input
         .lines()
         .map(|l| l.split(","))
-        .map(|s| s.map(|i| i.split("-").map(|j| j.parse::<i32>().unwrap()).collect_tuple::<(_, _)>().unwrap()).collect_vec())
+        .map(|s| s.map(|i| i.split("-").map(|j| j.parse::<u16>().unwrap()).collect_tuple::<(_, _)>().unwrap()).collect_vec())
         .filter(|r| overlaps(r[0], r[1]))
         .count();
 
     Solution::USize(overlaps_count)
 }
 
-fn fully_contains(a1: (i32, i32), a2: (i32, i32)) -> bool {
-    (a2.0 >= a1.0 && a2.1 <= a1.1) ||
-    (a1.0 >= a2.0 && a1.1 <= a2.1)
+fn fully_contains(a1: (u16, u16), a2: (u16, u16)) -> bool {
+    a2.0 >= a1.0 && a2.1 <= a1.1 || a1.0 >= a2.0 && a1.1 <= a2.1
 }
 
-fn overlaps(a1: (i32, i32), a2: (i32, i32)) -> bool {
+fn overlaps(a1: (u16, u16), a2: (u16, u16)) -> bool {
     a1.0 <= a2.1 && a2.0 <= a1.1
 }
